+++ src/main/java/github/anandb/netbeans/contract/PasteCallback.java
@@ -11,5 +11,4 @@ public interface PasteCallback {
     void onError(String message);
 
     void onAttachmentLimitReached();
-
 }